Support for interactive VM start without using tmux
If the global console setting is not set to tmux, `vm start -i contoso`
starts the VM in the background and launches cu(1) immediately to
connect to the console.

This feature helps avoid nested tmux sessions when the user is already
inside a tmux session. It is nearly equivalent to
`vm start contoso && vm console -w contoso`, but performs the same
action in a single step.