Refine cloud-init network-config creation
- Add IPv6 support (originally submitted by @6eh01der at #15)
- Use DHCP or RA when IP addresses were not specified in netconfig string
- Switch to routes from deprecated gateway[46]

In the previous implementation, when IP addresses were not specified,
for example, when only nameservers or searchdomains were provided, the IP
addresses remained unassigned. This PR changes the behavior to continue
obtaining IP addresses via DHCP or RA in such cases, allowing
nameservers, searchdomains, hostname and interface to be specified
alone.