Accept extra settings to pass to bhyve
bhyve can be dependent upon the order of the options it receives, so add
a setting similar to bhyve_options, called bhyve_extra, that is passed
to bhyve as the last options.

"-s 31,lpc" is passed by vm-bhyve after the options provided by
bhyve_options.  However, this causes bhyve to return an error when
attempting to use GPU passthru with an Intel iGPU due to additional
options needing to be passed.

Sample error:
$ bhyve ... -o pci.0.31.0.pcireg.vendor=host -s 31,lpc ...
pci slot 0:31:0 already occupied!